json: double conversion warning fix

+++ lib/utils/json.c
@@ -518,7 +518,7 @@ static int64_t decode_value(struct json_obj *obj,
 		*num = strtol(value->start, &endptr, 10);
 		if (*endptr == '.' && *(endptr + 1) >= '0' && *(endptr + 1) <= '9') {
              //This implementation supports just one decimal place
-			*num += (double)((*(endptr + 1) - '0') / 10.0);
+			*num += (float)((*(endptr + 1) - '0') / 10.0);
 		} else if (endptr != value->end) {
 			return -EINVAL;
 		}
